+++ webpack.config.js
@@ -12,12 +12,13 @@ import extCred from './.env.ssh.js';							/* Private credentials for ssh */
 // a. From wp-scripts
 import TerserPlugin from "terser-webpack-plugin";					/* already installed through WordPress scripts */
 import CopyPlugin from "copy-webpack-plugin";						/* already installed through WordPress scripts */
-import MiniCssExtractPlugin from "mini-css-extract-plugin";				/* already installed through WordPress scripts */
 // b. need the NPM package to be installed
 import CssMinimizerPlugin from "css-minimizer-webpack-plugin";
 import ImageMinimizerPlugin from "image-minimizer-webpack-plugin";
 import SSHWatchUploadWebpackPlugin from '@alexrah/ssh-watch-upload-webpack-plugin';	/* forked from ssh-watch-upload-webpack-plugin */
 import BrowserSyncPlugin from 'browser-sync-webpack-plugin';
+import RemoveEmptyScriptsPlugin from 'webpack-remove-empty-scripts';
+import * as sassTransform  from 'sass';
 
 // Utilities
 import { resolve, relative, dirname, join, parse } from 'path';
@@ -34,24 +35,28 @@ export default {
 	output: {
 	    path: resolve('./dist/'),
 	},
-	module: {
-		...wpConfig.module,
-		rules: [
-			...wpConfig.module.rules,
-		]
-	},
 	plugins: [
-		...wpConfig.plugins,
+
+		// Include WordPress plugins but filter out RTL plugin
+		...wpConfig.plugins.filter(
+			(plugin) => plugin.constructor.name !== 'RtlCssPlugin'
+		),
+		
+		// Remove empty JS files after WordPress asset generation
+		new RemoveEmptyScriptsPlugin({
+			stage: RemoveEmptyScriptsPlugin.STAGE_AFTER_PROCESS_PLUGINS,
+		}),
+		
 		// Runs only if "--watch" is passed in command-line
 		new BrowserSyncPlugin({
 			proxy: {
-			    target: extCred.proxy.address_http, /* must be in http, not in https, certif error otherwise */
-			    proxyReq: [
-				 function(proxyReq) {
-				 	// Allows to use lumiere codeception database
-					proxyReq.setHeader('X-Testing', 'true');
-				 }
-			    ],
+				target: extCred.proxy.address_http, /* must be in http, not in https, certif error otherwise */
+				proxyReq: [
+					function(proxyReq) {
+						// Allows to use lumiere codeception database
+						proxyReq.setHeader('X-Testing', 'true');
+					}
+				],
 			},
 			// Don't show any notifications in the browser
 			notify:true,
@@ -78,6 +83,7 @@ export default {
 		}),
 		new CopyPlugin( {
 			patterns: [
+			/****** All */
 			{
 				from: resolve( './src/' ),
 				globOptions: {
@@ -87,46 +93,27 @@ export default {
 						"**/vendor/bin/*",
 						"**/duck7000/imdb-graphql-php/doc/*",
 						"**/duck7000/imdb-graphql-php/src/Psr/**",
-						"**/vendor/twbs/bootstrap/build/*",
-						"**/vendor/twbs/bootstrap/js/**",
-						"**/vendor/twbs/bootstrap/nuget/*",
-						"**/vendor/twbs/bootstrap/scss/**",
-						"**/vendor/twbs/bootstrap/site/**",
-						"**/vendor/twbs/bootstrap/.github/**",
-						"**/vendor/twbs/bootstrap/package.js",
-						"**/vendor/twbs/bootstrap/hugo.yml",
-						"**/vendor/twbs/bootstrap/package.json",
-						"**/vendor/twbs/bootstrap/package-lock.json",
-						"**/vendor/twbs/bootstrap/.*",
-						"**/vendor/twbs/bootstrap/**/*.(map)",
+						"**/vendor/twbs/bootstrap/**",
 						"**/class/updates/.add_only_updates",
 					],
 				},
 			},
+			/****** Render files in block folders */
 			{
-				from: resolve( './src/assets/css/' ),
-				globOptions: {
-					concurrency: 100,
-				},
+				from: resolve( './src/assets/blocks/**/render.php' ),
 				to({ context, absoluteFilename }) {
-					/** add .min to filename */
-					return 'assets/css/[name].min[ext]';
-				},	
-			},
-			{
-				from: resolve( './src/assets/js/*.(js)' ),
-				globOptions: {
-					concurrency: 100,
+					const path = absoluteFilename.split("/").slice(7, -1).join("/");
+					return `${path}/[name][ext]`;
 				},
-				to({ context, absoluteFilename }) {
-					/** add .min to filename */
-					return 'assets/js/[name].min[ext]';
-				},	
 			},
+			/****** Bootstrap */
 			{
-				from: resolve( './src/assets/js/highslide/**/*.*' ),
+				from: resolve( './src/vendor/twbs/bootstrap/dist/**/*.(css|js)' ),
 				globOptions: {
 					concurrency: 100,
+					ignore: [
+						"**/vendor/twbs/bootstrap/dist/**/*[^min].(css|js)",
+					],
 				},
 				to({ context, absoluteFilename }) {
 					/**
@@ -140,6 +127,63 @@ export default {
 					return `${path}/[name][ext]`;
 				},
 			},
+			/****** CSS */
+			{
+				from: resolve( './src/assets/css/*.css' ),
+				globOptions: {
+					concurrency: 100,
+				},
+				to({ context, absoluteFilename }) {
+					/** add .min to filename */
+					return 'assets/css/[name].min.css';
+				},	
+			},
+			/****** SCSS */
+			{
+				from: resolve( './src/assets/**/*.scss' ),
+				globOptions: {
+					concurrency: 100,
+				},
+				to({ context, absoluteFilename }) {
+					const path = absoluteFilename.split("/").slice(7, -1).join("/");
+					/** add .min to filename */
+					return `${path}/[name].min.css`;
+				},
+				noErrorOnMissing: true,
+				// compile scss into css
+				transform: (content, path) => {
+					return sassTransform.compile(path).css
+				},	
+			},
+			/****** JS */
+			{
+				from: resolve( './src/assets/js/*.(js)' ),
+				globOptions: {
+					concurrency: 100,
+				},
+				to({ context, absoluteFilename }) {
+					/** add .min to filename */
+					return 'assets/js/[name].min[ext]';
+				},	
+			},
+			/****** Highslide JS */
+			{
+				from: resolve( './src/assets/js/highslide/**/*.*' ),
+				globOptions: {
+					concurrency: 100,
+				},
+				to({ context, absoluteFilename }) {
+					/**
+					 * @description Remove first & last item from ${path} array.
+					 * @example
+					 *      Orginal Path: 'src/images/avatar/image.jpg'
+					 *      Changed To: 'images/avatar'
+					 */
+					const path = absoluteFilename.split("/").slice(7, -1).join("/");
+					return `${path}/[name][ext]`;
+				},
+			},
+			/****** Pics */
 			{
 				from: resolve( './src/assets/pics/**/*.*' ),
 				globOptions: {
@@ -163,6 +207,12 @@ export default {
 		minimize: true,
 		minimizer: [
 			new TerserPlugin({
+				extractComments: false,  /* avoid creation of licence and other useless files in blocks */
+				terserOptions: {
+					format: {
+						comments: false,
+					},
+				},
 				parallel: 10,
 				test: /\.js$/i,
 				exclude: [ /assets\/js\/highslide\//, /vendor\// ],
@@ -178,6 +228,7 @@ export default {
 				},
 				parallel: 10,
 				test: /\.css$/i,
+				// Scss have already been renamed to .min.css
 				exclude: [ /vendor\// ],
 			}),
 			new ImageMinimizerPlugin({
@@ -195,5 +246,9 @@ export default {
 				exclude: [ /assets\/js\/highslide\//, /vendor\// ],
 			}),
 		],
-	}
+	},
+	
+	performance: {
+		maxAssetSize: 512000,
+	},
 };